requirements.txt
langchain
langchain-community
langchain-core
python-dotenv
colorama
langchain_groq 
ChatGroq
langchain-text-splitters
chromadb
tiktoken